// src/types.ts
export interface Note {
  [key: string]: unknown;
  id: number;
  content: string;
  deal_id: number | null;
  person_id: number | null;
  org_id: number | null;
  user_id: number;
  add_time: string;
  update_time: string;
  active_flag: boolean;
  pinned_to_deal_flag: boolean;
  pinned_to_person_flag: boolean;
  pinned_to_organization_flag: boolean;
  deal?: { title: string };
  person?: { name: string };
  org?: { name: string };
  user?: { name: string; email: string };
}